Reorganize views into platform-specific folders and extract shared components
